# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RestAPI for effective mobile",
        "title": "EffectiveMobile",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/subscriptions": {
            "get": {
                "description": "Get subscriptions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscritpion"
                ],
                "summary": "Get subscriptions",
                "operationId": "GetSubscriptions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SubscritpionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscritpion"
                ],
                "summary": "Create subscription",
                "operationId": "CreateSubscription",
                "parameters": [
                    {
                        "description": "Subscritpion creation data",
                        "name": "subscription",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.SubscritpionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SubscritpionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/subscriptions/sum": {
            "post": {
                "description": "Get subscriptions price sum",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscritpion"
                ],
                "summary": "Get subscriptions price sum",
                "operationId": "GetSubscriptionsPriceSum",
                "parameters": [
                    {
                        "description": "Filter",
                        "name": "filter",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.FilterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SumResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/subscriptions/{id}": {
            "get": {
                "description": "Get subscription by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscritpion"
                ],
                "summary": "Get subscription by id",
                "operationId": "GetSubscriptionByID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SubscritpionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Update subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscritpion"
                ],
                "summary": "Update subscription",
                "operationId": "UpdateSubscription",
                "parameters": [
                    {
                        "description": "Subscritpion update data",
                        "name": "subscription",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.SubscritpionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SubscritpionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Delete subscription",
                "tags": [
                    "Subscritpion"
                ],
                "summary": "Delete subscription",
                "operationId": "DeleteSubscritpion",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.FilterRequest": {
            "type": "object",
            "required": [
                "end_date",
                "start_date"
            ],
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.SubscritpionRequest": {
            "type": "object",
            "required": [
                "price",
                "service_name",
                "start_date",
                "user_id"
            ],
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.SubscritpionResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.SubscritpionsResponse": {
            "type": "object",
            "properties": {
                "subscriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SubscritpionResponse"
                    }
                }
            }
        },
        "dto.SumResponse": {
            "type": "object",
            "properties": {
                "sum": {
                    "type": "integer"
                }
            }
        }
    }
}